+++ src/main/java/be/baur/sdt/transform/XPathStatement.java
@@ -25,7 +25,7 @@ public abstract class XPathStatement extends Statement {
 	/**
 	 * Creates a statement from an XPath object.
 	 * 
-	 * @param xpath the XPath object, not null
+	 * @param xpath an XPath object, not null
 	 */
 	public XPathStatement(SDAXPath xpath) {
 		expression = Objects.requireNonNull(xpath, "xpath must not be null").toString();